+++ MongoFramework.Tests/DbEntityWriterTests.cs
@@ -1,20 +1,10 @@
 ﻿using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MongoFramework.Infrastructure;
+using MongoFramework.Tests.Models;
 
 namespace MongoFramework.Tests
 {
-	public class EntityWriterModel
-	{
-		public string Id { get; set; }
-		public string Title { get; set; }
-	}
-
-	public class ExtendedEntityWriterModel : EntityWriterModel
-	{
-		public string AdditionalField { get; set; }
-	}
-
 	[TestClass]
 	public class DbEntityWriterTests
 	{
