+++ MongoFramework.Tests/Models/ExtendedEntity.cs
@@ -1,9 +1,7 @@
 ﻿using System.Collections.Generic;
 
-namespace MongoFramework.Tests.Models
-{
-	class ExtendedEntity : CommonEntity
-	{
+namespace MongoFramework.Tests.Models {
+	internal class ExtendedEntity : CommonEntity {
 		public bool IsDisabled { get; set; }
 		public IDictionary<string, string> CustomData { get; set; }
 		public IEnumerable<string> Tags { get; set; }
