+++ MongoFramework.Tests/Models/ExtendedEntityWriterModel.cs
@@ -1,0 +1,6 @@
+﻿namespace MongoFramework.Tests.Models {
+	public class ExtendedEntityWriterModel : EntityWriterModel
+	{
+		public string AdditionalField { get; set; }
+	}
+}
